+++ app/modules/wallet-selector/ledger-wizard/flows.ts
@@ -3,7 +3,7 @@ import { toPairs, zip } from "lodash";
 import { pairZip } from "../../../../typings/modifications";
 import { GetState } from "../../../di/setupBindings";
 import { symbols } from "../../../di/symbols";
-import { injectableFn } from "../../../redux-injectify";
+import { injectableFn } from "../../../middlewares/redux-injectify";
 import { AppDispatch } from "../../../store";
 import { actions } from "../../actions";
 import {
