+++ app/modules/web3/flows.ts
@@ -1,5 +1,5 @@
 import { symbols } from "../../di/symbols";
-import { injectableFn } from "../../redux-injectify";
+import { injectableFn } from "../../middlewares/redux-injectify";
 import { AppDispatch } from "../../store";
 import { actions } from "../actions";
 import { NotificationCenter } from "../notifications/NotificationCenter";
