+++ test/integrationTestUtils.tsx
@@ -7,10 +7,10 @@ import { applyMiddleware, createStore, Store } from "redux";
 import { MemoryRouter } from "react-router";
 import { customizerContainerWithMiddlewareApi, setupBindings } from "../app/di/setupBindings";
 import { symbols } from "../app/di/symbols";
+import { createInjectMiddleware } from "../app/middlewares/redux-injectify";
 import { BrowserWalletConnector } from "../app/modules/web3/BrowserWallet";
 import { LedgerWalletConnector } from "../app/modules/web3/LedgerWallet";
 import { Web3Manager } from "../app/modules/web3/Web3Manager";
-import { createInjectMiddleware } from "../app/redux-injectify";
 import { IAppState, reducers } from "../app/store";
 import { InversifyProvider } from "../app/utils/InversifyProvider";
 import { dummyConfig, dummyLogger } from "./fixtures";
